requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
jsonschema>=4.23.0
openpyxl>=3.1.2

streamlit>=1.35.0
pandas>=2.2.0
pdfplumber>=0.11.0
python-docx>=1.1.2
google-generativeai>=0.7.2
pydantic>=2.7.0
openpyxl>=3.1.2

